Create smallest non-wearing window

Create an inactive period that represents the smallest non-wearing window. This function uses the make_inactive_period() function to create the non-wearing window.

Usage

make_smallest_nonwearing_window(min_non_wearing_length = 20 * 2)

Arguments

min_non_wearing_length

minimum non_wearing time before a bout is terminated

Value

An inactive period data frame that represents the smallest non-wearing window.

Examples

make_smallest_nonwearing_window()
#>    activity_counts bout non_wearing complete_day
#> 1                0    0        TRUE        FALSE
#> 2                0    0        TRUE        FALSE
#> 3                0    0        TRUE        FALSE
#> 4                0    0        TRUE        FALSE
#> 5                0    0        TRUE        FALSE
#> 6                0    0        TRUE        FALSE
#> 7                0    0        TRUE        FALSE
#> 8                0    0        TRUE        FALSE
#> 9                0    0        TRUE        FALSE
#> 10               0    0        TRUE        FALSE
#> 11               0    0        TRUE        FALSE
#> 12               0    0        TRUE        FALSE
#> 13               0    0        TRUE        FALSE
#> 14               0    0        TRUE        FALSE
#> 15               0    0        TRUE        FALSE
#> 16               0    0        TRUE        FALSE
#> 17               0    0        TRUE        FALSE
#> 18               0    0        TRUE        FALSE
#> 19               0    0        TRUE        FALSE
#> 20               0    0        TRUE        FALSE
#> 21               0    0        TRUE        FALSE
#> 22               0    0        TRUE        FALSE
#> 23               0    0        TRUE        FALSE
#> 24               0    0        TRUE        FALSE
#> 25               0    0        TRUE        FALSE
#> 26               0    0        TRUE        FALSE
#> 27               0    0        TRUE        FALSE
#> 28               0    0        TRUE        FALSE
#> 29               0    0        TRUE        FALSE
#> 30               0    0        TRUE        FALSE
#> 31               0    0        TRUE        FALSE
#> 32               0    0        TRUE        FALSE
#> 33               0    0        TRUE        FALSE
#> 34               0    0        TRUE        FALSE
#> 35               0    0        TRUE        FALSE
#> 36               0    0        TRUE        FALSE
#> 37               0    0        TRUE        FALSE
#> 38               0    0        TRUE        FALSE
#> 39               0    0        TRUE        FALSE
#> 40               0    0        TRUE        FALSE