Generate the smallest bout with the smallest non-wearing period dataset

This function creates a dataset consisting of the smallest bout and the smallest non-wearing period. The bout length, non-wearing period length, and epoch length are defined in the global variables: minimum_bout_length, maximum_number_consec_inactive_epochs_in_bout, and min_non_wearing_length, respectively.

Usage

make_smallest_bout_with_smallest_non_wearing_period()

Value

A data frame with columns for activity counts and date-time stamps.

Examples

make_smallest_bout_with_smallest_non_wearing_period()
#>    activity_counts bout non_wearing complete_day                time
#> 1                0    0       FALSE        FALSE 2012-04-07 00:00:30
#> 2                0    0       FALSE        FALSE 2012-04-07 00:01:00
#> 3                0    0       FALSE        FALSE 2012-04-07 00:01:30
#> 4                0    0       FALSE        FALSE 2012-04-07 00:02:00
#> 5              500    1       FALSE        FALSE 2012-04-07 00:02:30
#> 6              500    1       FALSE        FALSE 2012-04-07 00:03:00
#> 7              500    1       FALSE        FALSE 2012-04-07 00:03:30
#> 8              500    1       FALSE        FALSE 2012-04-07 00:04:00
#> 9              500    1       FALSE        FALSE 2012-04-07 00:04:30
#> 10             500    1       FALSE        FALSE 2012-04-07 00:05:00
#> 11             500    1       FALSE        FALSE 2012-04-07 00:05:30
#> 12             500    1       FALSE        FALSE 2012-04-07 00:06:00
#> 13             500    1       FALSE        FALSE 2012-04-07 00:06:30
#> 14             500    1       FALSE        FALSE 2012-04-07 00:07:00
#> 15               0    0        TRUE        FALSE 2012-04-07 00:07:30
#> 16               0    0        TRUE        FALSE 2012-04-07 00:08:00
#> 17               0    0        TRUE        FALSE 2012-04-07 00:08:30
#> 18               0    0        TRUE        FALSE 2012-04-07 00:09:00
#> 19               0    0        TRUE        FALSE 2012-04-07 00:09:30
#> 20               0    0        TRUE        FALSE 2012-04-07 00:10:00
#> 21               0    0        TRUE        FALSE 2012-04-07 00:10:30
#> 22               0    0        TRUE        FALSE 2012-04-07 00:11:00
#> 23               0    0        TRUE        FALSE 2012-04-07 00:11:30
#> 24               0    0        TRUE        FALSE 2012-04-07 00:12:00
#> 25               0    0        TRUE        FALSE 2012-04-07 00:12:30
#> 26               0    0        TRUE        FALSE 2012-04-07 00:13:00
#> 27               0    0        TRUE        FALSE 2012-04-07 00:13:30
#> 28               0    0        TRUE        FALSE 2012-04-07 00:14:00
#> 29               0    0        TRUE        FALSE 2012-04-07 00:14:30
#> 30               0    0        TRUE        FALSE 2012-04-07 00:15:00
#> 31               0    0        TRUE        FALSE 2012-04-07 00:15:30
#> 32               0    0        TRUE        FALSE 2012-04-07 00:16:00
#> 33               0    0        TRUE        FALSE 2012-04-07 00:16:30
#> 34               0    0        TRUE        FALSE 2012-04-07 00:17:00
#> 35               0    0        TRUE        FALSE 2012-04-07 00:17:30
#> 36               0    0        TRUE        FALSE 2012-04-07 00:18:00
#> 37               0    0        TRUE        FALSE 2012-04-07 00:18:30
#> 38               0    0        TRUE        FALSE 2012-04-07 00:19:00
#> 39               0    0        TRUE        FALSE 2012-04-07 00:19:30
#> 40               0    0        TRUE        FALSE 2012-04-07 00:20:00
#> 41               0    0        TRUE        FALSE 2012-04-07 00:20:30
#> 42               0    0        TRUE        FALSE 2012-04-07 00:21:00
#> 43               0    0        TRUE        FALSE 2012-04-07 00:21:30
#> 44               0    0        TRUE        FALSE 2012-04-07 00:22:00
#> 45               0    0        TRUE        FALSE 2012-04-07 00:22:30
#> 46               0    0        TRUE        FALSE 2012-04-07 00:23:00
#> 47               0    0        TRUE        FALSE 2012-04-07 00:23:30
#> 48               0    0        TRUE        FALSE 2012-04-07 00:24:00
#> 49               0    0        TRUE        FALSE 2012-04-07 00:24:30
#> 50               0    0        TRUE        FALSE 2012-04-07 00:25:00
#> 51               0    0        TRUE        FALSE 2012-04-07 00:25:30
#> 52               0    0        TRUE        FALSE 2012-04-07 00:26:00
#> 53               0    0        TRUE        FALSE 2012-04-07 00:26:30
#> 54               0    0        TRUE        FALSE 2012-04-07 00:27:00